+++ NEWS.md
@@ -39,6 +39,7 @@
 * Fixed lint_package bug where cache was not caching (#146, @schloerke)
 * Fixed cache not saved in a directory other than requested (#213, @fangly)
 * Commas linter handles missing arguments calls properly (#145)
+* Add `function_left_parentheses_linter` to check that there is no space between a function name and its left parentheses (#204, @jrnold).
 
 # lintr 1.0.0 #
 * bugfix to work with testthat 1.0.0